Jump Resist Begone
Removes being able to use jump to resist out of grabs.
It's that simple. Makes captures obscenely painful if someone knows about this. Removed elsewhere, and while it looks like someone tried to 'fix this' via a cooldown? It's not great. At all. Period. Effectively an exploit. IS an exploit in some use cases. I'm near certain the intent isn't to allow someone to drop from chain grabs.
Outta here.